Replace "angularjs" playground framework with "css"
The legacy "angularjs" framework causes an error because ux-aspects-ng1.js is no longer provided. Since all AngularJS content has now been removed, the legacy playgrounds should be HTML+CSS only. The `CssPlaygroundStrategy` therefore replaces the `AngularJSPlaygroundStrategy` for legacy content, and only includes CSS dependencies.
